+++ sjf4j/src/test/java/org/sjf4j/SimpleExampleTest.java
@@ -99,7 +99,7 @@ public class SimpleExampleTest {
         jo.put("extra", "blabla");
         // See also: `putNonNull()`, `putIfAbsent()`, `computeIfAbsent()`
 
-        jo.toBuilder().putIfAbsent("x", "xx").put("y", "yy");
+        jo.edit().putIfAbsent("x", "xx").put("y", "yy");
         // Supports Builder-style chained operations
 
         jo.remove("extra");